Stabilize JavaFX/Scala runtime wiring and restore drawing UX.
Align Scala runtime dependencies to Scala 3, make interpreter loading robust under IDE classpaths, restore Prettify/hover behavior, and add decimal support only for TEXT-AT coordinates without affecting shape rendering.

+++ src/main/scala/drawing/model/Command.scala
@@ -6,6 +6,6 @@ case class BoundingBox(x1: Int, y1: Int, x2: Int, y2: Int) extends Command
 case class Line(x1: Int, y1: Int, x2: Int, y2: Int) extends Command
 case class Rectangle(x1: Int, y1: Int, x2: Int, y2: Int) extends Command
 case class Circle(x: Int, y: Int, radius: Int) extends Command
-case class TextAt(x: Int, y: Int, text: String) extends Command
+case class TextAt(x: Double, y: Double, text: String) extends Command
 case class Draw(color: Color, items: List[Command]) extends Command
 case class Fill(color: Color, item: Command) extends Command
